# docs/openapi.yml
openapi: 3.1.0
info:
  title: Java API Security Challenge - Core Criteria
  description: Core Criteria for the Java Security Exercise
  version: 0.2.3
  contact:
    name: Video Games

servers:
  - url: http://localhost:4000
tags:
  - name: games
paths:
  /games:
    post:
      tags:
        - games
      summary: Create a game
      description: Create a new game.
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGame'
      responses:
        '201':
          description: Successfully created a new Game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
    get:
      tags:
        - games
      summary: Get all games
      description: Get a list of every game.
      operationId: getGames
      responses:
        '200':
          description: Get a list of all games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllGames'
  /games/{id}:
    put:
      tags:
        - games
      summary: Update a game
      description: Update an existing game. For ease of implementation, all fields are required from the client.
      operationId: updateGame
      parameters:
        - name: id
          in: path
          description: The ID of the game to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGame'
      responses:
        '201':
          description: Successfully updated the specified game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
    delete:
      tags:
        - games
      summary: Delete a game
      description: Delete an existing game. When deleting data, it's useful to send the deleted record back to the client so they can re-create it if deletion was a mistake.
      operationId: deleteGame
      parameters:
        - name: id
          in: path
          description: The ID of the game to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully deleted the specified game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'

components:
  schemas:
    CreateGame:
      type: object
      required: [title, age_rating, number_of_players]
      properties:
        title:
          type: string
          example: "League of Legends"
        age_rating:
          type: string
          example: "10+"
        number_of_players:
          type: integer
          example: 10
    AllGames:
      type: array
      items:
        $ref: '#/components/schemas/Game'
    Game:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "League of Legends"
        age_rating:
          type: string
          example: "10+"
        number_of_players:
          type: integer
          example: 10